pyproject.toml
[project]
name = "mcp-server-hubspot"
version = "0.1.0"
description = "A simple Hubspot MCP server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["mcp>=1.4.1", "hubspot-api-client>=11.1.0", "python-dotenv>=1.0.1", "faiss-cpu>=1.7.4", "numpy>=1.24.0", "sentence-transformers>=2.2.2", "huggingface-hub==0.14.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["pyright>=1.1.389"]

[project.scripts]
mcp-server-hubspot = "mcp_server_hubspot:run_main"